Improve handling for too recent log backups
If a too recent log backup is found, the search date is adjusted and it will look for earlier backups.  Its possible that the modified date on the file we need got adjusted and it has a later modified date than the file that is too recent.
If we found an earlier log backup, then log an error and continue checking the headers of later backups.