Optimize iterations
Iterations now add work to a queue (channel).  If a single database is taking a long time to process, the next loop can now begin with that database skipped as it's already in progress.  If a single database needs to catch up, it will no longer impact other databases as long as sufficient threads are available.